+++ src/App.tsx
@@ -1,7 +1,13 @@
 import RecipeEntry from "./components/recipe/RecipeEntry";
+import RecipeList from "./components/recipe/RecipeList";
 
 function App() {
-  return <RecipeEntry />;
+  return (
+    <>
+      <RecipeEntry />
+      <RecipeList />
+    </>
+  );
 }
 
 export default App;